new cart

+++ lib/modules/shop_laptobs_app/Shopdashboardadmin/laptop/addlap.dart
@@ -321,7 +321,7 @@ class addlap extends StatelessWidget {
                                   lapname: nameController.text,
                                   lapdesc: descreptionController.text,
                                   Ram: ramval,
-                                  price: double.parse(priceController.text),
+                                  price: int.parse(priceController.text),
                                   hardsize: hardval,
                                   brand: brandval,
                                   hardtype: hardtypeval);
@@ -332,7 +332,7 @@ class addlap extends StatelessWidget {
                                   lapname: nameController.text,
                                   lapdesc: descreptionController.text,
                                   Ram: ramval,
-                                  price: double.parse(priceController.text),
+                                  price: int.parse(priceController.text),
                                   hardsize: hardval,
                                   brand: brandval,
                                   hardtype: hardtypeval);
